Issue #23 MikroTik webfig encryption/authentication added. MikroTik also got moved into a subdirectory within protocol
