+++ .golangci.yml
@@ -3,8 +3,8 @@ run:
   timeout: 1m
   skip-dirs:
   skip-files:
-   - protocol/mikrotik.go
-   - protocol/mikrotik_test.go
+   - protocol/mikrotik/mikrotik_test.go
+   - protocol/mikrotik/msg.go
   build-tags:
 linters:
 # https://golangci-lint.run/usage/linters/
@@ -42,7 +42,7 @@ linters:
     - gocritic
     - gocyclo
     - godot
-    #- godox
+    - godox
     - goerr113
     - gofmt
     - gofumpt
@@ -73,7 +73,7 @@ linters:
     - tenv
     - unconvert
     - unparam
-    - usestdlibvars
+    #- usestdlibvars
     #- varnamelen
     - wastedassign
     - whitespace
