+++ go.mod
@@ -5,5 +5,6 @@ go 1.19
 require (
 	github.com/lor00x/goldap v0.0.0-20180618054307-a546dffdd1a3
 	github.com/vjeantet/ldapserver v1.0.1
-	golang.org/x/text v0.9.0
+	golang.org/x/crypto v0.11.0
+	golang.org/x/text v0.11.0
 )
