+++ protocol/mikrotik/mikrotik_test.go
@@ -1,4 +1,4 @@
-package protocol
+package mikrotik
 
 import (
 	"bytes"
@@ -98,7 +98,7 @@ func Test_CVE_2018_14847_ServerResponse2(t *testing.T) {
 		t.Error("Raw var 3 was expected to contain 0xbc bytes, has", len(result.Raw[3]))
 	}
 
-	if 0 != bytes.Compare(result.Raw[3], []byte("\x5a\x00\x4d\x32\x10\x00\x00\xa8\x00\x00\x1c\x00\x00\x00\x0a\x00\xfe\x00\x05\x00\x00\x09\x00\x06\x00\x00\x09\x00\x0b\x00\x00\x08\xfe\xff\x07\x00\x12\x00\x00\x09\x02\x01\x00\xfe\x09\x01\x02\x00\x00\x09\x03\x09\x00\xfe\x21\x13\x73\x79\x73\x74\x65\x6d\x20\x64\x65\x66\x61\x75\x6c\x74\x20\x75\x73\x65\x72\x11\x00\x00\x21\x00\x01\x00\x00\x21\x05\x61\x64\x6d\x69\x6e\x62\x00\x4d\x32\x10\x00\x00\xa8\x00\x00\x1c\x00\x00\x00\x0a\x00\xfe\x00\x05\x00\x00\x09\x00\x06\x00\x00\x09\x00\x1f\x00\x00\x08\x7e\x9e\x07\x00\x0b\x00\x00\x08\xfe\xff\x07\x00\x12\x00\x00\x09\x02\x01\x00\xfe\x09\x01\x02\x00\x00\x09\x03\x09\x00\xfe\x21\x13\x73\x79\x73\x74\x65\x6d\x20\x64\x65\x66\x61\x75\x6c\x74\x20\x75\x73\x65\x72\x11\x00\x00\x21\x00\x01\x00\x00\x21\x05\x61\x64\x6d\x69\x6e")) {
+	if !bytes.Equal(result.Raw[3], []byte("\x5a\x00\x4d\x32\x10\x00\x00\xa8\x00\x00\x1c\x00\x00\x00\x0a\x00\xfe\x00\x05\x00\x00\x09\x00\x06\x00\x00\x09\x00\x0b\x00\x00\x08\xfe\xff\x07\x00\x12\x00\x00\x09\x02\x01\x00\xfe\x09\x01\x02\x00\x00\x09\x03\x09\x00\xfe\x21\x13\x73\x79\x73\x74\x65\x6d\x20\x64\x65\x66\x61\x75\x6c\x74\x20\x75\x73\x65\x72\x11\x00\x00\x21\x00\x01\x00\x00\x21\x05\x61\x64\x6d\x69\x6e\x62\x00\x4d\x32\x10\x00\x00\xa8\x00\x00\x1c\x00\x00\x00\x0a\x00\xfe\x00\x05\x00\x00\x09\x00\x06\x00\x00\x09\x00\x1f\x00\x00\x08\x7e\x9e\x07\x00\x0b\x00\x00\x08\xfe\xff\x07\x00\x12\x00\x00\x09\x02\x01\x00\xfe\x09\x01\x02\x00\x00\x09\x03\x09\x00\xfe\x21\x13\x73\x79\x73\x74\x65\x6d\x20\x64\x65\x66\x61\x75\x6c\x74\x20\x75\x73\x65\x72\x11\x00\x00\x21\x00\x01\x00\x00\x21\x05\x61\x64\x6d\x69\x6e")) {
 		t.Error("Raw var 3 does not contain the expected payload")
 	}
 }
@@ -124,7 +124,7 @@ func Test_MessageCreation(t *testing.T) {
 	data := msg.Serialize()
 
 	readit := NewM2Message()
-	if !ParseM2Message([]byte(data), readit) {
+	if !ParseM2Message(data, readit) {
 		t.Error("Failed M2 messaging parsing")
 	}
 }
